SEQ grand rework with C++20

+++ tasks/melnik_i_min_neigh_diff_vec/common/include/common.hpp
@@ -7,7 +7,7 @@
 
 namespace melnik_i_min_neigh_diff_vec {
 
-using InType = std::vector<double>;
+using InType = std::vector<int>;
 using OutType = std::tuple<int, int>;
 using TestType = std::tuple<int, std::string>; //TODO rework in future
 using BaseTask = ppc::task::Task<InType, OutType>;
